Refactoring REST server
